+++ ICD.Connect.Settings/Cores/ICoreSettings.cs
@@ -1,4 +1,6 @@
-﻿using ICD.Connect.Settings.Header;
+﻿using System.Collections.Generic;
+using ICD.Connect.Settings.Header;
+using ICD.Connect.Settings.Validation;
 
 namespace ICD.Connect.Settings.Cores
 {
@@ -15,5 +17,11 @@ namespace ICD.Connect.Settings.Cores
 		/// <param name="config"></param>
 		/// <returns></returns>
 		ConfigurationHeader GetHeader(string config);
+
+		/// <summary>
+		/// Validates the core settings as a whole.
+		/// </summary>
+		/// <returns></returns>
+		IEnumerable<SettingsValidationResult> Validate();
 	}
 }